feat: timer 시작/정지 api 모듈 연결
